Move code to mutualize CgmesIidmMapping
