+++ cgmes/cgmes-extensions/src/main/java/com/powsybl/cgmes/extensions/CgmesIidmMapping.java
@@ -112,8 +112,6 @@ public interface CgmesIidmMapping extends Extension<Network> {
 
     Set<CgmesTopologicalNode> getUnmappedTopologicalNodes();
 
-    void invalidateTopology();
-
     Map<Double, BaseVoltageSource> getBaseVoltages();
 
     BaseVoltageSource getBaseVoltage(double nominalVoltage);
